feat: implement checkout and order history (UC-08)
- Order and OrderItem models to represent a placed order and its line items
- OrderDAO with methods to persist orders and items to the AivenDB
- OrderService to handle checkout logic, payment processing, and order retrieval
- CheckoutScreen with delivery/collection toggle, address input, and payment flow
- OrderConfirmationScreen to display their order summary after successful checkout
- OrderHistoryScreen to list with expandable item details

+++ src/main/java/com/novasolutions/ipospu/impl/PU_PaymentAPI.java
@@ -31,7 +31,7 @@ public class PU_PaymentAPI implements PaymentAPI {
      * @return true if authorised, false if declined or service unavailable
      */
     @Override
-    public boolean processPayment(double amount, int cardNumber, String expiry) {
+    public boolean processPayment(double amount, long cardNumber, String expiry) {
         if (amount <= 0) return false;
         if (!isValidExpiry(expiry)) return false;
